refactor: Convert Search page, api mocks  to TypeScript implementation
