+++ src/hooks/tests/util/index.tsx
@@ -1,5 +1,9 @@
 import { QueryClient, QueryClientProvider } from "react-query";
 
+interface Props {
+  children: React.ReactNode;
+}
+
 export const createWrapper = (retry = true) => {
   const queryClient = new QueryClient({
     defaultOptions: {
@@ -9,7 +13,7 @@ export const createWrapper = (retry = true) => {
     },
   });
 
-  return ({ children }) => (
+  return ({ children }: Props) => (
     <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
   );
 };
